[modify] region_group_id str -> int 수정

+++ app/persistence/model/region_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, SmallInteger
 
 from app import db
 from app.extensions.utils.time_helper import get_server_timestamp
@@ -8,7 +8,7 @@ from core.domains.region.entity.region_entity import RegionEntity
 class RegionModel(db.Model):
     __tablename__ = "regions"
 
-    id = Column(Integer().with_variant(Integer, "sqlite"), primary_key=True)
+    id = Column(SmallInteger().with_variant(Integer, "sqlite"), primary_key=True)
     name = Column(String(length=50), nullable=False)
     created_at = Column(DateTime, default=get_server_timestamp())
     updated_at = Column(DateTime, default=get_server_timestamp())
